DOCSP-10446: feedback fixes for bulkwrite
